Fixed LifeBar Implemented new LifeBar Deleted old LifeBar Added template progress bar to the prefabs Did some formatting

+++ UnityProject/Assets/Scripts/Game/Testing/LifebarTester.cs
@@ -1,22 +1,28 @@
 using Game.UI;
 using UnityEngine;
 
-namespace Game.Testing {
-	public class LifebarTester : MonoBehaviour {
-		[SerializeField]                  private int maxLifePoints;
-		[SerializeField, Range(0f, 100f)] private float lifePoints;
+namespace Game.Testing
+{
+	public class LifebarTester : MonoBehaviour
+	{
+		[SerializeField]                private int maxLifePoints;
+		[SerializeField, Range(0, 100)] private int lifePoints;
 
-		private void Update() {
+		private void Update()
+		{
 			UpdateLifeBar();
 		}
 
-		public void UpdateLifeBar() {
+		public void UpdateLifeBar()
+		{
 			LifeBar.Instance.MaxPoints = maxLifePoints;
-			
+
 			var points = LifeBar.Instance.Points;
-			if (points < lifePoints) {
+			if (points < lifePoints)
+			{
 				LifeBar.Instance.AddPoints(lifePoints - points);
-			} else if (points > lifePoints) {
+			} else if (points > lifePoints)
+			{
 				LifeBar.Instance.RemovePoints(points - lifePoints);
 			}
 		}
